requirements.txt
streamlit==1.34.0
python-dotenv==1.0.1
streamlit-option-menu==0.3.12